tracer-master/requirements.txt
appdirs==1.4.3
BeautifulSoup==3.2.1
beautifulsoup4==4.5.3
bs4==0.0.1
certifi==2017.1.23
dj-database-url==0.4.2
Django==1.11
django-cors-headers==2.0.2
djangorestframework==3.6.2
gunicorn==19.7.1
heroku==0.1.4
lxml==3.7.3
oauthlib==2.0.2
packaging==16.8
pprint==0.1
psycopg2==2.7.1
pycurl==7.43.0
pyparsing==2.2.0
python-dateutil==1.5
pytz==2017.2
requests==2.13.0
requests-oauthlib==0.8.0
selenium==3.3.3
six==1.10.0
tweepy==3.5.0
urllib3==1.20
virtualenv==15.1.0
waitress==1.0.2
wfastcgi==3.0.0
whitenoise==3.3.0
